ai-service/requirements.txt
fastapi
uvicorn
pydantic
requests
beautifulsoup4
torch
transformers
ftfy
timm
clip-anytorch
Pillow
